Moved CaptureOutput to "test" package for reuse
